fix: MSVC bring-up round 10 -- binary-mode writers, time_t max check, localtime macro removal
Completes the round-9/10 failure census (69 fails at run 29053499592), all
root-caused statically from the CI log:

1. CRLF corruption (AccountManagement.RestoresLegacyFiles*/MigrationCleansUp*/
   RestoredSnapshot*, PlayerFinalize.ByteIdenticalAndSingleFile): Windows CRT
   text mode expanded '\n' to "\r\n" inside byte-exact payloads, turning
   restored legacy files' historical "\r\n" endings into "\r\r\n".
   open_secure_output_file (account_management.cpp) gains _O_BINARY + "wb"
   (same mapping db.cpp's open_secure_temp_output_file already uses);
   write_player_text (db.cpp) opens its scratch "wb" so the pinned player-text
   serialization keeps identical bytes across platforms; the PlayerFinalize
   test's own write_file helper writes "wb" (byte-identity is the assertion).

2. AccountManagement.FormatsAccountSummaries*/FormatsPendingVerification*:
   format_account_timestamp's overflow guard did
   static_cast<long>(numeric_limits<time_t>::max()) -- on LLP64 that truncates
   INT64_MAX to -1, so every positive timestamp rendered "Invalid". Compare in
   long long instead.

3. DamageMethodTest RTC3 (ClampsNonAmbushOverflowDamage/ShieldAbsorption):
   Crash_load's num_of_hours was read by the pre-charge cost expression before
   its only assignment -- historical UB (uninitialized read + potential
   divide-by-zero) that MSVC's runtime checks turn into 0xc0000409. Initialize
   to 0; the live rent paths overwrite it before any behavioral use.

4. Remove structs.h's '#define localtime(x) localtime((time_t*)x)' cast macro
   -- the enabler that hid every long*/time_t* mismatch f6ec28c fixed. All call
   sites now pass a real time_t*, so future mismatches fail MSVC compilation
   instead of corrupting stack reads at runtime.

macOS gate: build + 644/0 ctest (71 skips) + boot-golden verify clean.
32-bit container gate: build + 644/0 ctest.

+++ src/account_management_presentation.cpp
@@ -5,7 +5,11 @@ std::string format_account_timestamp(long timestamp)
     if (timestamp <= 0)
         return "Never";
 
-    if (timestamp > static_cast<long>(std::numeric_limits<time_t>::max()))
+    // Compare in long long: casting time_t's max down to long truncates to -1 on
+    // Windows LLP64 (long = 4 bytes, time_t = 8), which made every positive
+    // timestamp "Invalid" (Phase 3 Task 6). Both operands fit in long long on
+    // every supported platform.
+    if (static_cast<long long>(timestamp) > static_cast<long long>(std::numeric_limits<time_t>::max()))
         return "Invalid";
 
     time_t raw_time = static_cast<time_t>(timestamp);
